pyproject.toml
[tool.poetry]
name = "nonebot-plugin-add-friends"
version = "0.1.3"
description = "处理好友申请和群邀请，支持查看申请、手动同意/拒绝申请、同意/拒绝全部申请。"
authors = ["hakunomiko <48401273+hakunomiko@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "nonebot_plugin_add_friends"}]
repository = "https://github.com/hakunomiko/nonebot-plugin-add-friends"
keywords = ["nonebot", "nonebot2", "plugin"]

[tool.poetry.dependencies]
python = ">=3.9"
nonebot2 = "^2.3.0"
nonebot-adapter-onebot = ">=2.4.4"
nonebot_plugin_localstore = "^0.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
line-length = 120
select = [
    "F",
    "E",
    "W",
    "I",
    "UP",
    "RUF"
]
ignore = [
    "E402",   # module-import-not-at-top-of-file
    "UP006",  # non-pep585-annotation
    "UP035",  # deprecated-import
    "UP037",  # quoted-annotation
    "RUF001", # ambiguous-unicode-character-string
    "RUF002", # ambiguous-unicode-character-docstring
    "RUF003", # ambiguous-unicode-character-comment
]
